Fake out some UI things from all tests
