+++ test/test_helpers.rb
@@ -7,6 +7,7 @@ module TestHelpers
   autoload :FakeContext, 'test_helpers/fake_context'
   autoload :FakeFS, 'test_helpers/fake_fs'
   autoload :FakeProject, 'test_helpers/fake_project'
+  autoload :FakeUI, 'test_helpers/fake_ui'
   autoload :Project, 'test_helpers/project'
   autoload :Schema, 'test_helpers/schema'
 end
